requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
bitarray==2.9.2
certifi==2024.2.2
charset-normalizer==3.3.2
ckzg==1.0.1
click==8.1.7
colorama==0.4.6
cytoolz==0.12.3
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.1.0
eth-utils==4.1.0
eth_abi==5.1.0
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.0.5
nodeenv==1.8.0
parsimonious==0.10.0
prisma==0.13.1
protobuf==5.26.1
pycryptodome==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
pyunormalize==15.1.0
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
rlp==4.0.0
rpds-py==0.18.0
sniffio==1.3.1
StrEnum==0.4.15
tomlkit==0.12.4
toolz==0.12.1
typing_extensions==4.11.0
urllib3==2.2.1
web3==6.17.1
websockets==12.0
yarl==1.9.4
